Trying to fix windows compile

+++ Engine/Common/DataStoring/DataStore.h
@@ -97,7 +97,7 @@ namespace Leviathan{
 
 		DLLEXPORT void AddVar(NamedVariableList* newvaluetoadd);
 		DLLEXPORT void AddVar(std::shared_ptr<NamedVariableList> values);
-		DLLEXPORT void Remove(unsigned int index);
+		DLLEXPORT void Remove(size_t index);
 		DLLEXPORT void Remove(const std::string &name);
 
 		DLLEXPORT int GetVariableType(const std::string &name) const;
