+++ Engine/Entities/CommonStateObjects.h
@@ -179,7 +179,7 @@ namespace Leviathan{
     class TrackControllerState : public ObjectDeltaStateData{
     public:
 
-        DLLEXPORT TrackControllerState(int tick, int reached, float speed, float progress);
+        DLLEXPORT TrackControllerState(int tick, size_t reached, float speed, float progress);
         DLLEXPORT TrackControllerState(int tick, sf::Packet &packet);
             
         DLLEXPORT virtual void CreateUpdatePacket(ObjectDeltaStateData* olderstate,
